Add auth logic, final changes for first release

+++ server/api/task/delete.post.ts
@@ -9,10 +9,20 @@ const bodySchema = z.object({
 export default defineEventHandler(async (e) => {
   checkAPIEnabled()
   
+  const userId = await getUserId(e)
   const bodyParse = await readValidatedBody(e, (b) => bodySchema.safeParse(b))
   const bodyData = checkParseResult(bodyParse)
 
   const db = useDB(e)
+  
+  const boardInfo = await getBoardInfo(db, bodyData.boardId, userId)
+  if (!canEdit(boardInfo.isOwner, boardInfo.publicPerms)) {
+    throw createError({
+      statusCode: 400,
+      message: "You do not have permission to delete tasks on this board."
+    })
+  }
+
   await checkTaskExists(db, bodyData.boardId, bodyData.taskId)
 
   await db.deleteTask(bodyData.taskId)
